♻️ Modifier GRD raccordement
